refactor auth serivce with dedicated secure service,refactor test
