--- server/github/index.ts
@@ -1,2 +1,0 @@
-export * from './githubsync.service'
-export * from './githubauth.service'
