+++ server/github/services/index.ts
@@ -1,0 +1,3 @@
+export { GithubAuthService } from './githubauth.service'
+export { Secure } from './secure.service'
+export { GithubSyncService } from './githubsync.service'
