+++ server/github/test/githubauth.service.test.ts
@@ -1,15 +1,17 @@
 import { User } from "db";
 import { Context, createMockContext, MockContext } from "db/test";
+import { setupServer, SetupServerApi } from 'msw/node';
 import { Octokit } from "octokit";
-import { GithubAuthService } from "server/github";
-import { setupServer, SetupServerApi } from 'msw/node'
-import { commitsHandler, languagesHandler, rateLimitNotOkHandler, rateLimitOkHandler, subscriptionsHandler, userHandler } from "./githubApiServerStubHandler";
-import { Secure } from "../githubauth.service";
+import { GithubAuthService } from "server/github/services";
+import { Secure } from "../services/secure.service";
+import { userHandler } from "./githubApiServerStubHandler";
 
-jest.mock('../githubauth.service')
+jest.mock('../services/secure.service')
+
 
 const secureMock = Secure as jest.MockedClass<typeof Secure>
-const myApi = new Secure() as jest.Mocked<Secure>;
+// const secureMock: jest.Mocked<Secure> = new Secure() as jest.Mocked<Secure>;
+// const secureMock: jest.Mocked<Secure> = { encrypt: jest.fn(), decrypt: jest.fn() }
 
 let worker: SetupServerApi
 let mockCtx: MockContext
@@ -29,58 +31,18 @@ afterEach(() => {
 })
 
 describe('GithubAuthService', () => {
-    it('should encrypt value when APP_HASH_SECRET env is set', () => {
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
+    it('should not save OAuth user to db when token param is not set', async () => {
+        const service: GithubAuthService = new GithubAuthService(ctx.prisma, secureMock.prototype)
 
-        const result: string = service.encrypt(tobeencrypted)
-
-        expect(result).toContain('.')
-    });
-    it('should throw exception when APP_HASH_SECRET env is not set', () => {
-
-        delete process.env.APP_HASH_SECRET
-
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
-
-        expect(service.encrypt).toThrow()
-    });
-    it('should throw exception when value parameter is empty', () => {
-
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
-
-        expect(() => service.encrypt('')).toThrow()
-    });
-    it('should decrypt when env and parameters are valid', () => {
-
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
-
-        const decrypted = service.decrypt(encrypted);
-
-        expect(decrypted).toBe(tobeencrypted)
-    });
-    it('should throw exception when APP_HASH_SECRET env is not set', () => {
-
-        delete process.env.APP_HASH_SECRET
-
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
-
-        expect(service.decrypt).toThrow()
-    });
-    it('should throw exception when value parameter is empty', () => {
-
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
-
-        expect(() => service.decrypt('')).toThrow()
-    });
-    it('should throw exception when value parameter does not contains dot', () => {
-
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
-
-        expect(() => service.decrypt('8c223ea342d96de2353608c482160e82')).toThrow()
+        await expect(service.secureThenSaveOauthUser('', 'fake@mail.com', 'fakename')).rejects.toThrow();
     });
     it('should save OAuth user to db when params are correct', async () => {
 
-        const userDbMock: User = {
+        secureMock.prototype.encrypt.mockImplementationOnce((value) => {
+            return encrypted
+        })
+
+        const userDbStub: User = {
             id: 1000,
             hashedPassword: encrypted,
             email: 'fake@mail.com',
@@ -90,28 +52,60 @@ describe('GithubAuthService', () => {
             role: ''
         }
 
-        mockCtx.prisma.user.upsert.mockResolvedValue(userDbMock);
+        mockCtx.prisma.user.upsert.mockResolvedValue(userDbStub);
 
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
+        const service: GithubAuthService = new GithubAuthService(ctx.prisma, secureMock.prototype)
 
         const userDb = await service.secureThenSaveOauthUser(tobeencrypted, 'fake@mail.com', 'fakename');
 
         expect(mockCtx.prisma.user.upsert).toBeCalledTimes(1)
 
-        expect(userDb.id).toBe(userDbMock.id)
+        expect(secureMock.prototype.encrypt).toHaveBeenCalledTimes(1);
+
+        expect(userDb.id).toBe(userDbStub.id)
 
     });
-    it('should ', async () => {
+    it('should authenticate against to Github when sessionId correspond to OAuth user', async () => {
 
         worker = setupServer(userHandler())
 
         worker.listen()
 
-        const service: GithubAuthService = new GithubAuthService(ctx.prisma)
+        secureMock.prototype.decrypt.mockImplementationOnce((value) => {
+            return 'stubbed'
+        })
 
-        const lo = secureMock.mock.instances[0]!
+        const userDbStub: User = {
+            id: 1,
+            hashedPassword: encrypted,
+            email: 'fake@mail.com',
+            name: 'fakename',
+            createdAt: new Date(),
+            updatedAt: new Date(),
+            role: ''
+        }
 
+        mockCtx.prisma.user.findFirst.mockResolvedValueOnce(userDbStub);
 
-        const authenticated = await service.authenticateGithubBySessionId(1);
+        const service: GithubAuthService = new GithubAuthService(ctx.prisma, secureMock.prototype)
+
+        await expect(service.authenticateGithubBySessionId(1)).resolves.toBe(true);
+
+        expect(secureMock.prototype.decrypt).toHaveBeenCalledTimes(1);
+    });
+
+    it('should not authenticate against to Github when sessionId not correspond to OAuth user', async () => {
+
+        worker = setupServer(userHandler())
+
+        worker.listen()
+
+        const userDbStub: User | null = null
+
+        mockCtx.prisma.user.findFirst.mockResolvedValueOnce(userDbStub);
+
+        const service: GithubAuthService = new GithubAuthService(ctx.prisma, secureMock.prototype)
+
+        await expect(service.authenticateGithubBySessionId(1)).resolves.toBe(false);
     });
 });
