+++ server/github/test/githubsync.service.test.ts
@@ -1,17 +1,15 @@
+import { GithubRepo, GithubRepoLanguage, GithubRepoTopics } from 'db';
+import { Context, createMockContext, MockContext } from 'db/test';
+import fs from 'fs';
+import { setupServer, SetupServerApi } from 'msw/node';
 import { Octokit } from "octokit";
-import { GithubSyncService, GithubSyncServiceGenerator } from 'server/github';
-import { closeGithubApiServerStub, resetHandlersServerStub, listenGithubApiServerStub } from "./githubApiServerStub";
-import { createMockContext, Context, MockContext, prismaMockSingleton } from 'db/test'
-import { GithubRepoLanguage, GithubRepoTopics, PrismaClient } from 'db'
-import fs from 'fs'
-import path from 'path'
+import path from 'path';
+import { GhApiCommitConverter } from "server/github/converters";
+import { GithubSyncService } from 'server/github/services';
+import { repoDbToDto } from "shared/mappers";
+import { IGithubDto } from "shared/models";
 import { GhApiRepoConvert } from "../converters/GhApiRepoConvert";
 import { IGhApiAggregate, IGhAPICommit, IGhApiRepo } from "../githubapi.model";
-import { repoApiToDto, repoDbToDto } from "shared/mappers";
-import { GithubRepo } from 'db'
-import { IGithubDto } from "shared/models";
-import { GhApiCommitConverter } from "server/github/converters";
-import { setupServer, SetupServerApi } from 'msw/node'
 import { commitsHandler, languagesHandler, rateLimitNotOkHandler, rateLimitOkHandler, subscriptionsHandler, userHandler } from "./githubApiServerStubHandler";
 
 let worker: SetupServerApi
@@ -24,8 +22,6 @@ let apiLanguageList: { [key: string]: number }
 const client = new Octokit({ auth: 'aSuperFakeTokenStubbedByMsw' })
 
 beforeAll(() => {
-
-    // worker.listen()
 
     apiResponseSubscriptions = fs.readFileSync(path.join(__dirname, './fixtures/api.github.com.user.subscriptions.json'), 'utf-8')
